+++ SVD/common/matrix_funcs.c
