+++ SVD/serial_mpi_implementation/results/svd_mpi_6M.log
